+++ src/components/Planning/PlanningClassesTable/PlanningClassesRow/index.ts
@@ -1,0 +1,1 @@
+export { default as PlanningClassesRow } from './PlanningClassesRow';
