--- src/components/Planning/PlanningClassesTable/PlanningClassesTableRow/index.ts
@@ -1,1 +1,0 @@
-export { default as PlanningClassesTableRow } from './PlanningClassesTableRow';
